requirements-simple.txt
fastapi
uvicorn
pydantic
httpx
beautifulsoup4
feedparser
sqlalchemy
python-dotenv
pyjwt
aiohttp
jinja2
lxml
pytest
python-multipart
redis
newspaper3k
cachetools
alembic 